Разметка TobaccoEditDialog и логика работы с полями вынесена в отдельный файл
