+++ src/components/Editors/TobaccoEditor/TobaccoEditorProps.ts
@@ -1,0 +1,7 @@
+import { TobaccoClass } from "../../../Classes";
+
+export interface TobaccoEditorProps {
+  tobaccoData: TobaccoClass;
+  setNewTobaccosData: (tobacco: TobaccoClass) => void;
+  pullNewPhoto: (file: File) => void;
+}
