+++ src/components/Editors/index.ts
@@ -1,0 +1,1 @@
+export { TobaccoEditor } from "./TobaccoEditor/TobaccoEditor";
